int8_t instead of int

+++ to edit/world/block.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <stdint.h>
+
 class Block
 {
 public:
-	int nr;
+	int8_t nr;
 	int x, y;
 	
 	Block();
